+++ data/duke.txt
@@ -5,3 +5,4 @@
 [T][ ]gaysex
 [D][ ]gaysex  (by: Sep 8 2021 5pm)
 [E][ ]orgy  (at: Aug 7 2022 7pm)
+[T][ ]sex
