+++ src/main/java/Duke/Duke.java
@@ -1,9 +1,11 @@
+package Duke;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Objects;
 import java.util.Scanner;
 import java.io.File;
 import java.io.FileWriter;
+
 
 public class Duke {
     public static void main(String[] args) {
