+++ src/main/java/Duke/DukeException.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class DukeException extends Exception{
     public DukeException(String message) {
         super(message);
