+++ src/main/java/Duke/Parser.java
@@ -1,5 +1,8 @@
+
+package Duke;
 import java.util.ArrayList;
 import java.util.Scanner;
+
 
 public class Parser {
