+++ src/main/java/Duke/Storage.java
@@ -1,8 +1,11 @@
+package Duke;
+
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
 import java.io.File;
 import java.io.FileWriter;
+
 
 
 public class Storage {
