+++ src/main/java/Duke/Todo.java
@@ -1,3 +1,6 @@
+
+package Duke;
+
 public class Todo extends Task{
